Move Navmesh code away from dPhysics
