+++ CMakeLists.txt
@@ -14,7 +14,7 @@ string(REPLACE "\n" ";" variables ${variables})
 foreach(variable ${variables})
 	# If the string contains a #, skip it
 	if(NOT "${variable}" MATCHES "#")
-	
+
 		# Split the variable into name and value
 		string(REPLACE "=" ";" variable ${variable})
 
@@ -43,7 +43,7 @@ set(PROJECT_VERSION "${PROJECT_VERSION_MAJOR}.${PROJECT_VERSION_MINOR}.${PROJECT
 message(STATUS "Version: ${PROJECT_VERSION}")
 
 # Compiler flags:
-# Disabled deprecated warnings as the MySQL includes have deprecated code in them. 
+# Disabled deprecated warnings as the MySQL includes have deprecated code in them.
 # Disabled misleading indentation as DL_LinkedList from RakNet has a weird indent.
 # Disabled no-register
 # Disabled unknown pragmas because Linux doesn't understand Windows pragmas.
@@ -118,17 +118,18 @@ endforeach()
 set(INCLUDED_DIRECTORIES
 	"dCommon"
 	"dChatFilter"
-	"dGame" 
-	"dGame/dBehaviors" 
-	"dGame/dComponents" 
-	"dGame/dGameMessages" 
-	"dGame/dInventory" 
-	"dGame/dMission" 
-	"dGame/dEntity" 
-	"dGame/dUtilities" 
-	"dPhysics" 
-	"dZoneManager" 
-	"dDatabase" 
+	"dGame"
+	"dGame/dBehaviors"
+	"dGame/dComponents"
+	"dGame/dGameMessages"
+	"dGame/dInventory"
+	"dGame/dMission"
+	"dGame/dEntity"
+	"dGame/dUtilities"
+	"dPhysics"
+	"dNavigation"
+	"dZoneManager"
+	"dDatabase"
 	"dDatabase/Tables"
 	"dNet"
 	"dScripts"
@@ -192,7 +193,7 @@ file(
 file(
 	GLOB HEADERS_DGAME
 	LIST_DIRECTORIES false
-	${PROJECT_SOURCE_DIR}/dGame/Entity.h 
+	${PROJECT_SOURCE_DIR}/dGame/Entity.h
 	${PROJECT_SOURCE_DIR}/dGame/dGameMessages/GameMessages.h
 	${PROJECT_SOURCE_DIR}/dGame/EntityManager.h
 	${PROJECT_SOURCE_DIR}/dScripts/CppScripts.h
@@ -206,6 +207,7 @@ add_subdirectory(dNet)
 add_subdirectory(dScripts) # Add for dGame to use
 add_subdirectory(dGame)
 add_subdirectory(dZoneManager)
+add_subdirectory(dNavigation)
 add_subdirectory(dPhysics)
 
 # Create a list of common libraries shared between all binaries
@@ -230,18 +232,18 @@ add_subdirectory(dMasterServer) # Add MasterServer last so it can rely on the ot
 
 # Add our precompiled headers
 target_precompile_headers(
-	dGame PRIVATE 
+	dGame PRIVATE
 	${HEADERS_DGAME}
 )
 
 target_precompile_headers(
-	dZoneManager PRIVATE 
+	dZoneManager PRIVATE
 	${HEADERS_DZONEMANAGER}
 )
 
 # Need to specify to use the CXX compiler language here or else we get errors including <string>.
 target_precompile_headers(
-	dDatabase PRIVATE 
+	dDatabase PRIVATE
 	"$<$<COMPILE_LANGUAGE:CXX>:${HEADERS_DDATABASE}>"
 )
