+++ dNavigation/CMakeLists.txt
@@ -1,0 +1,3 @@
+set(DNAVIGATION_SOURCES "dNavMesh.cpp")
+
+add_library(dNavigation STATIC ${DNAVIGATION_SOURCES})
