+++ dNavigation/DetourExtensions.h
@@ -1,0 +1,25 @@
+#pragma once
+
+#include "Recast.h"
+#include "DetourCommon.h"
+#include "DetourNavMesh.h"
+#include "DetourNavMeshBuilder.h"
+#include "DetourNavMeshQuery.h"
+
+static const int NAVMESHSET_MAGIC = 'M' << 24 | 'S' << 16 | 'E' << 8 | 'T'; // char[4] of 'MSET'
+static const int NAVMESHSET_VERSION = 1;
+
+struct NavMeshSetHeader {
+	int magic;
+	int version;
+	int numTiles;
+	dtNavMeshParams params;
+};
+
+struct NavMeshTileHeader {
+	dtTileRef tileRef;
+	int dataSize;
+};
+
+static const int MAX_POLYS = 256;
+static const int MAX_SMOOTH = 2048;
